Clean and organize run index code
