+++ graphrag/index/run/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2024 Microsoft Corporation.
+# Licensed under the MIT License
+
+"""Run module for GraphRAG."""
+
+from graphrag.index.run.run import run_pipeline, run_pipeline_with_config
+
+__all__ = ["run_pipeline", "run_pipeline_with_config"]
